Refactors lyft service into two services: one for omniauth and one for making api calls.  Also refactors tests for these services into two files
