+++ spec/models/lyft_service_spec.rb
@@ -1,50 +1,8 @@
 require 'rails_helper'
 
 describe 'Lyft Service' do
-  describe '#profile_info' do
-    it 'returns a parsed JSON response with user profile information' do
-      user = create(:user)
-
-      stub_request(:get, "https://api.lyft.com/v1/profile").
-           to_return(status: 200, body: {
-                "id": "123456789",
-                "first_name": "Rick",
-                "last_name": "Sanchez",
-                "has_taken_a_ride": true
-            }.to_json, headers: {})
-
-      lyft_service = LyftService.new(user.lyft_token, user.lyft_refresh_token)
-      actual = lyft_service.profile_info
-
-      expect(actual).to be_a(Hash)
-      expect(actual[:id]).to eq("123456789")
-      expect(actual[:first_name]).to eq("Rick")
-      expect(actual[:last_name]).to eq("Sanchez")
-    end
-  end
-
-  describe '#renew_token' do
-    it 'returns an active lyft token' do
-      user = create(:user)
-
-      stub_request(:post, 'https://api.lyft.com/oauth/token').
-           to_return(status: 200, body: {
-             access_token: 'jnuf9348fnci98w3rendoirfo3in4coi',
-             token_type: 'bearer',
-             expires_in: 3600,
-             scope: 'profile offline rides.read public rides.request'
-            }.to_json, headers: {})
-
-      lyft_service = LyftService.new(user.lyft_token, user.lyft_refresh_token)
-      actual = lyft_service.renew_token
-
-      expect(actual).to be_a(String)
-      expect(actual).to eq('jnuf9348fnci98w3rendoirfo3in4coi')
-    end
-  end
-
   describe '#call_ride' do
-    it 'returns a ride status of pending' do
+    it 'calls ride and returns ride information' do
       user = create(:user)
       origin = { lat: 37.77663, lng: -122.39227 }
       destination = { lat: 37.771, lng: -122.39123 }
@@ -74,7 +32,7 @@ describe 'Lyft Service' do
                 }
               }.to_json, headers: {})
 
-      lyft_service = LyftService.new(user.lyft_token, user.lyft_refresh_token)
+      lyft_service = LyftService.new(user)
       actual = lyft_service.call_ride(origin, destination)
 
       expect(actual).to be_a(String)
@@ -88,13 +46,14 @@ describe 'Lyft Service' do
       origin = { lat: 37.77663, lng: -122.39227 }
       destination = { lat: 37.771, lng: -122.39123 }
 
-      stub_request(:get, 'https://api.lyft.com/v1/cost?end_lat=37.771&end_lng=-122.39123&start_lat=37.77663&start_lng=-122.39227&ride_type=lyft').
-        with(headers: {
-                        'Accept'=>'*/*',
-                        'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3',
-                        'Authorization'=>"Bearer #{user.lyft_token}",
-                        'User-Agent'=>'Faraday v0.12.2'
-                         }).
+      stub_request(:get, "https://api.lyft.com/v1/cost?end_lat=37.771&end_lng=-122.39123&ride_type=lyft&start_lat=37.77663&start_lng=-122.39227").
+                     with(
+                       headers: {
+                      'Accept'=>'*/*',
+                      'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3',
+                      'Authorization'=>'Bearer',
+                      'User-Agent'=>'Faraday v0.12.2'
+                       }).
         to_return(status: 200, body: '{
                                         "cost_estimates": [
                                             {
@@ -113,9 +72,9 @@ describe 'Lyft Service' do
                                                 "can_request_ride": true
                                             }
                                         ]
-                                    }')
+                                    }', headers:{})
 
-      lyft_service = LyftService.new(user.lyft_token, user.lyft_refresh_token)
+      lyft_service = LyftService.new(user)
       actual = lyft_service.get_estimate(origin, destination)
 
       expect(actual[:min_cost]).to eq(1052)
